+++ tests/aac_144.v
@@ -1,0 +1,10 @@
+From Coq Require Import ZArith.
+From AAC_tactics Require Import AAC.
+
+Goal forall X:Prop, X->X.
+Succeed (try aac_rewrite Z.gcd_mod).
+Abort.
+
+Goal forall X:Prop, X->X.
+Succeed (try aac_normalise).
+Abort.